moved tools from cmd/ to tools/
