+++ hack/cherry-pick.sh
@@ -32,7 +32,7 @@ if [[ ! -d "${UPSTREAM_REPO_LOCATION}" ]]; then
 fi
 
 if [[ -z "${NO_REBASE-}" ]]; then
-  lastrev="$(go run ${OS_ROOT}/hack/version.go ${OS_ROOT}/Godeps/Godeps.json ${repo}/${package})"
+  lastrev="$(go run ${OS_ROOT}/tools/godepversion/godepversion.go ${OS_ROOT}/Godeps/Godeps.json ${repo}/${package})"
 fi
 
 pushd "${UPSTREAM_REPO_LOCATION}" > /dev/null
