+++ hack/update-generated-docs.sh
@@ -9,7 +9,7 @@ set -o pipefail
 OS_ROOT=$(dirname "${BASH_SOURCE}")/..
 source "${OS_ROOT}/hack/common.sh"
 
-"${OS_ROOT}/hack/build-go.sh" cmd/gendocs
+"${OS_ROOT}/hack/build-go.sh" tools/gendocs
 
 # Find binary
 gendocs="$(os::build::find-binary gendocs)"
@@ -19,7 +19,7 @@ if [[ -z "$gendocs" ]]; then
     echo "It looks as if you don't have a compiled gendocs binary"
     echo
     echo "If you are running from a clone of the git repo, please run"
-    echo "'./hack/build-go.sh cmd/gendocs'."
+    echo "'./hack/build-go.sh tools/gendocs'."
   } >&2
   exit 1
 fi
