+++ hack/verify-upstream-commits.sh
@@ -7,7 +7,7 @@ set -o pipefail
 OS_ROOT=$(dirname "${BASH_SOURCE}")/..
 source "${OS_ROOT}/hack/common.sh"
 
-"${OS_ROOT}/hack/build-go.sh" cmd/rebasehelpers/commitchecker
+"${OS_ROOT}/hack/build-go.sh" tools/rebasehelpers/commitchecker
 
 # Find binary
 commitchecker="$(os::build::find-binary commitchecker)"
