+++ tools/genbashcomp/gen_openshift_bash_comp.go
