+++ tools/gendocs/gen_openshift_docs.go
